model_deploy/cats_and_dogs/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
torch==2.5.0
Pillow==9.5.0
torchvision==0.20.0
